requirements.txt
crewai
crewai-tools
openai
pydantic
pypdf
requests
beautifulsoup4
python-dotenv
